+++ cookbooks/bcpc_kafka/CHANGELOG.md
@@ -1,11 +1,11 @@
-kafka-bcpc CHANGELOG
+bcpc_kafka CHANGELOG
 ====================
 
-This file is used to list changes made in each version of the kafka-bcpc cookbook.
+This file is used to list changes made in each version of the bcpc_kafka cookbook.
 
 0.1.0
 -----
-- [Bloomberg Finance L.P.] - Initial release of kafka-bcpc
+- [Bloomberg Finance L.P.] - Initial release of bcpc_kafka
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.